Refactor LED HW blinking and bootloader / reset effect handling
Currently we read the states of LED pins from LAN switch, WAN PHY and
MiniPCIe ports in the main loop, and after reading, we blink the
corresponding RGB LEDs accordingly, also in the main loop.

We also have a timer handler for RGB LED effect/pattern, with is run
with 15 Hz frequency and is used for bootloader blinking pattern and
knight rider reset initialization pattern.

Refactor the code so that:

* Reading the states of HW LED pins is left in the main loop, but is
  refactored so that all pin states are stored into one uint32_t
  variable (called input_led_pins). We also store the states of LED pins
  that were previously not read at all (the second LEDs of LAN pins, and
  LED pins of port 5 from the switch).

* The RGB LED pattern timer handler is run with 1000 Hz frequency
  instead of 15 Hz, and also handles blinking of RGB LEDs according to
  states from input_led_pins, if reset initialization pattern is not
  active. Moving the RGB LED blinking code from the main loop to this
  handler makes the main loop faster by around 40%. The downside is that
  although the LED pin states are read more often, their changes are
  propagated to RGB LEDs only once per millisecond. This should not be a
  problem, because the devices (LAN switch, WAN PHY, and MiniPCIe /
  mSATA cards) shouldn't blink the LED's that fast: indeed, the LAN
  switch minimum LED blink rate is 21 ms.

We also set the priority of the I2C slave interrupt to come after the
LED pattern interrupt.

Signed-off-by: Marek Behún <marek.behun@nic.cz>