power_control: Read MANRES instead of SYSRES_OUT pin when selecting reset type
We keep reading SYSRES_OUT_PIN in power_control_first_startup() to
determine whether the manual reset button is pressed in order to select
reset type the user wants to trigger.

But according to Armada 385 datasheet, the SYSRST_OUTn is asserted low
for an additional 100 ms after manual reset is deasserted.

This makes the code think that the button is pressed for a little while,
even if it is not pressed, resulting in the first RGB LED blinking a
green color (the start of the reset selector LED pattern) before the
knight rider initialization pattern takes over.

Fix this by reading the MANRES pin instead. Afterwards wait until
SYSRES_OUT is deasserted.

Signed-off-by: Marek Behún <marek.behun@nic.cz>