i2c_iface: Fix cmd_get_features
When cmd_get_features is given an additional argument and the features
are to be read from flash, always fail is flashing is busy.

When the function is called with cmd_len == 2, clear the reply in case
we fail, so that we won't send the reply from the previous call (when
cmd_len == 1).

Signed-off-by: Marek Behún <marek.behun@nic.cz>