i2c_iface: Fix setting button pressed interrupt when setting user mode
When setting the front button to user mode where previously it was not
in user mode, we want to set the button pressed interrupt if the button
is pressed. But the code to do that is currently never executed, because
we check for previous mode in a variable that we already set to current
mode (there is redundant setting of button.user_mode). Fix this.

Signed-off-by: Marek Behún <kabel@kernel.org>